Combination optimization beta release
First release of the functioning combination optimization tools. Sentiment analysis is not implemented as of now due to runtime concerns. This project is intended for use in local environments where computational resources are not ideal for computing optimal weights for enormous  combination spaces. Sentiment analysis will be implemented at a later stage with optimized pipelines to externally deliver the sentiment input for each component.